# Lazily initialize exponential histogram buckets