--- a/sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/internal/aggregator/DoubleExponentialHistogramAggregatorTest.java
+++ b/sdk/metrics/src/test/java/io/opentelemetry/sdk/metrics/internal/aggregator/DoubleExponentialHistogramAggregatorTest.java
@@ -57,13 +57,14 @@ class DoubleExponentialHistogramAggregatorTest {
       InstrumentationScopeInfo.empty();
   private static final MetricDescriptor METRIC_DESCRIPTOR =
       MetricDescriptor.create("name", "description", "unit");
+  private static final ExponentialBucketStrategy EXPONENTIAL_BUCKET_STRATEGY =
+      ExponentialBucketStrategy.newStrategy(160, ExponentialCounterFactory.mapCounter());
 
   private static Stream<DoubleExponentialHistogramAggregator> provideAggregator() {
     return Stream.of(
         aggregator,
         new DoubleExponentialHistogramAggregator(
-            ExemplarReservoir::doubleNoSamples,
-            ExponentialBucketStrategy.newStrategy(160, ExponentialCounterFactory.mapCounter())));
+            ExemplarReservoir::doubleNoSamples, EXPONENTIAL_BUCKET_STRATEGY));
   }
 
   private static int valueToIndex(int scale, double value) {
@@ -83,8 +84,19 @@ class DoubleExponentialHistogramAggregatorTest {
 
   @Test
   void createHandle() {
-    assertThat(aggregator.createHandle())
-        .isInstanceOf(DoubleExponentialHistogramAggregator.Handle.class);
+    AggregatorHandle<?, ?> handle = aggregator.createHandle();
+    assertThat(handle).isInstanceOf(DoubleExponentialHistogramAggregator.Handle.class);
+    ExponentialHistogramAccumulation accumulation =
+        ((DoubleExponentialHistogramAggregator.Handle) handle)
+            .doAccumulateThenReset(Collections.emptyList());
+    assertThat(accumulation.getPositiveBuckets())
+        .isInstanceOf(DoubleExponentialHistogramAggregator.EmptyExponentialHistogramBuckets.class);
+    assertThat(accumulation.getPositiveBuckets().getScale())
+        .isEqualTo(EXPONENTIAL_BUCKET_STRATEGY.getStartingScale());
+    assertThat(accumulation.getNegativeBuckets())
+        .isInstanceOf(DoubleExponentialHistogramAggregator.EmptyExponentialHistogramBuckets.class);
+    assertThat(accumulation.getNegativeBuckets().getScale())
+        .isEqualTo(EXPONENTIAL_BUCKET_STRATEGY.getStartingScale());
   }
 
   @Test
@@ -108,6 +120,8 @@ class DoubleExponentialHistogramAggregatorTest {
     int expectedScale = 5; // should be downscaled from 20 to 5 after recordings
 
     assertThat(acc.getScale()).isEqualTo(expectedScale);
+    assertThat(acc.getPositiveBuckets().getScale()).isEqualTo(expectedScale);
+    assertThat(acc.getNegativeBuckets().getScale()).isEqualTo(expectedScale);
     assertThat(acc.getZeroCount()).isEqualTo(2);
 
     // Assert positive recordings are at correct index
@@ -163,6 +177,8 @@ class DoubleExponentialHistogramAggregatorTest {
 
     // With 160 buckets allowed, minimum scale is -4
     assertThat(acc.getScale()).isEqualTo(-4);
+    assertThat(acc.getPositiveBuckets().getScale()).isEqualTo(-4);
+    assertThat(acc.getNegativeBuckets().getScale()).isEqualTo(-4);
 
     // if scale is -4, base is 65,536.
     int base = 65_536;
@@ -370,8 +386,11 @@ class DoubleExponentialHistogramAggregatorTest {
       d += min;
     }
 
-    ExponentialHistogramAccumulation acc = handle.accumulateThenReset(Attributes.empty());
-    assertThat(Objects.requireNonNull(acc).getScale()).isEqualTo(3);
+    ExponentialHistogramAccumulation acc =
+        Objects.requireNonNull(handle.accumulateThenReset(Attributes.empty()));
+    assertThat(acc.getScale()).isEqualTo(3);
+    assertThat(acc.getPositiveBuckets().getScale()).isEqualTo(3);
+    assertThat(acc.getNegativeBuckets().getScale()).isEqualTo(3);
     assertThat(acc.getPositiveBuckets().getBucketCounts().size()).isEqualTo(160);
     assertThat(acc.getPositiveBuckets().getTotalCount()).isEqualTo(n);
   }
@@ -380,17 +399,22 @@ class DoubleExponentialHistogramAggregatorTest {
   void testDownScale() {
     DoubleExponentialHistogramAggregator.Handle handle =
         (DoubleExponentialHistogramAggregator.Handle) aggregator.createHandle();
+    // record a measurement to initialize positive buckets
+    handle.recordDouble(0.5);
+
     handle.downScale(20); // down to zero scale
 
     // test histogram operates properly after being manually scaled down to 0
-    handle.recordDouble(0.5);
     handle.recordDouble(1.0);
     handle.recordDouble(2.0);
     handle.recordDouble(4.0);
     handle.recordDouble(16.0);
 
-    ExponentialHistogramAccumulation acc = handle.accumulateThenReset(Attributes.empty());
-    assertThat(Objects.requireNonNull(acc).getScale()).isEqualTo(0);
+    ExponentialHistogramAccumulation acc =
+        Objects.requireNonNull(handle.accumulateThenReset(Attributes.empty()));
+    assertThat(acc.getScale()).isEqualTo(0);
+    assertThat(acc.getPositiveBuckets().getScale()).isEqualTo(0);
+    assertThat(acc.getNegativeBuckets().getScale()).isEqualTo(0);
     ExponentialHistogramBuckets buckets = acc.getPositiveBuckets();
     assertThat(acc.getSum()).isEqualTo(23.5);
     assertThat(buckets.getOffset()).isEqualTo(-2);
@@ -516,6 +540,8 @@ class DoubleExponentialHistogramAggregatorTest {
     assertThat(acc.getZeroCount()).isEqualTo(numberOfUpdates);
     assertThat(acc.getSum()).isCloseTo(100.0D * 10000, Offset.offset(0.0001)); // float error
     assertThat(acc.getScale()).isEqualTo(3);
+    assertThat(acc.getPositiveBuckets().getScale()).isEqualTo(3);
+    assertThat(acc.getNegativeBuckets().getScale()).isEqualTo(3);
     MetricAssertions.assertThat(acc.getPositiveBuckets())
         .hasTotalCount(numberOfUpdates * 3)
         .hasOffset(-27);