Exponential histograms separately track bucket counts in the negative and positive range. Currently, histograms are not permitted to record negative measurements, yet we allocate memory to negative buckets anyway. The exponential histogram aggregation logic works with negative measurements in anticipation of them being permitted in the future, but even when they are permitted, many instruments will only record positive values. 

This PR adjusts the exponential histogram aggregation to lazily instantiate both positive and negative buckets.

Here are the results of the `HistogramCollectBenchmark` before:
```
Benchmark                                                                          (aggregationGenerator)  (aggregationTemporality)  Mode  Cnt           Score           Error   Units
HistogramCollectBenchmark.recordAndCollect                                   EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5  9197564316.600 ± 157971347.306   ns/op
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate                    EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           3.764 ±         0.040  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate.norm               EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5    38305473.600 ±    816439.674    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space           EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           3.949 ±        15.903  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space.norm      EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5    40265318.400 ± 162152456.251    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Survivor_Space       EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5          ≈ 10⁻³                  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Survivor_Space.norm  EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5        3412.800 ±     29385.237    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.count                         EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           3.000                  counts
HistogramCollectBenchmark.recordAndCollect:·gc.time                          EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           5.000                      ms
HistogramCollectBenchmark.recordAndCollect                                   EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5  9216800908.200 ±  31120256.517   ns/op
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate                    EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           3.806 ±         0.089  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate.norm               EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5    38807947.200 ±    830507.303    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space           EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           3.783 ±        15.757  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space.norm      EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5    38587596.800 ± 160860384.216    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.count                         EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           3.000                  counts
HistogramCollectBenchmark.recordAndCollect:·gc.time                          EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           7.000                      ms
```
And after:
```
HistogramCollectBenchmark.recordAndCollect                               EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5  9363457049.800 ± 348621966.366   ns/op
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate                EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           3.069 ±         0.125  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate.norm           EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5    31767908.800 ±    760387.947    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space       EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           1.946 ±        10.261  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space.norm  EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5    20132659.200 ± 106153700.699    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.count                     EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           2.000                  counts
HistogramCollectBenchmark.recordAndCollect:·gc.time                      EXPONENTIAL_BUCKET_HISTOGRAM                     DELTA    ss    5           6.000                      ms
HistogramCollectBenchmark.recordAndCollect                               EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5  9282971800.200 ± 643932624.558   ns/op
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate                EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           3.136 ±         0.119  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.alloc.rate.norm           EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5    32196094.400 ±    868793.468    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space       EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           1.939 ±        10.229  MB/sec
HistogramCollectBenchmark.recordAndCollect:·gc.churn.G1_Eden_Space.norm  EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5    20132659.200 ± 106153700.699    B/op
HistogramCollectBenchmark.recordAndCollect:·gc.count                     EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           2.000                  counts
HistogramCollectBenchmark.recordAndCollect:·gc.time                      EXPONENTIAL_BUCKET_HISTOGRAM                CUMULATIVE    ss    5           5.000                      ms
```
`gc.alloc.rate` improves from 3.764 MB/sec to 3.069 MB/sec (~18% reduction) for delta temporality, and 3.806 MB/sec to 3.136 MB/sec (~18%) for cumulative temporality.